Integer to float grade.coefficient and grade.grade

+++ database/migrations/2018_04_29_140221_grade.php
@@ -15,8 +15,8 @@ class Grade extends Migration
     {
         Schema::create('grade', function (Blueprint $table) {
             $table->increments('grade_id');
-            $table->integer('grade');
-            $table->integer('coefficient');
+            $table->float('grade');
+            $table->float('coefficient');
             $table->integer('subject_id')->unsigned();
             $table->foreign('subject_id')->references('subject_id')->on('subject');
             $table->integer('student_id')->unsigned();
